moved azure folder
