+++ pkg/providers/azure/keyvault/auth.go
