+++ pkg/providers/azure/keyvault/logger.go
